상품 등록 예외처리 추가

+++ src/main/java/com/idle/shoppingmall/Service/Product/ProductService.java
@@ -16,7 +16,7 @@ public class ProductService {
     private final ProductMapper productMapper;
 
     @Transactional
-    public Long addProduct(Product productDTO) {
+    public Integer addProduct(Product productDTO) {
         return productMapper.saveProduct(productDTO);
     } // addProduct
